ai_service/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
numpy
Pillow
manga-ocr
easyocr
deep-translator
requests